- (snmp_api.h, snmp_api.c, snmp_auth.c, snmp_impl.h, asn1.c): 64bit cleanup.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@3049 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/snmp_api.h
@@ -79,7 +79,7 @@ struct snmp_pdu {
 	/*
 	 * Protocol-version independent fields
 	 */
-    int	    version;
+    long    version;
     int	    command;	/* Type of this PDU */
     long    reqid;	/* Request id - note: not incremented on retries */
     long    msgid;      /* Message id for V3 messages
